Automatically backed up by Learn

+++ index.js
@@ -17,11 +17,11 @@ function divide(a,b) {
 
 
 function inc(n) {
-  return parseInt(n++)
+  return parseFloat(n++)
 }
 
 function dec(n) {
-  return parseInt(n--)
+  return parseFloat(n--)
 }
 
 function makeInt(n) {
